Extract Monaco editor creation into separate file
