+++ client/src/languages/javascript.ts
@@ -1,12 +1,11 @@
-import * as monaco from 'monaco-editor';
-import {h,createProjector,VNode} from 'maquette';
+import {h,VNode} from 'maquette';
 import * as Langs from '../definitions/languages'; 
 import * as Graph from '../definitions/graph'; 
 import * as Values from '../definitions/values'; 
+import {createEditor} from '../editors/editor';
 
 import ts from 'typescript';
 import axios from 'axios';
-
 import {Md5} from 'ts-md5';
 
 declare var PYTHONSERVICE_URI: string;
@@ -111,57 +110,7 @@ class JavascriptBlockKind implements Langs.Block {
           [ (cell.code.value==undefined) ? evalButton : ("Value is: " + JSON.stringify(cell.code.value)) ]),
       ]);
  
-      let afterCreateHandler = (el) => { 
-        let ed = monaco.editor.create(el, {
-          value: state.block.source,
-          language: 'javascript',
-          scrollBeyondLastLine: false,
-          theme:'vs',
-          minimap: { enabled: false },
-          overviewRulerLanes: 0,
-          lineDecorationsWidth: "0ch",
-          fontSize: 14,
-          fontFamily: 'Monaco',
-          lineNumbersMinChars: 2,
-          lineHeight: 20,
-          lineNumbers: "on",
-          scrollbar: {
-            verticalHasArrows: true,
-            horizontalHasArrows: true,
-            vertical: 'none',
-            horizontal: 'none'
-          }
-        });    
-
-        ed.createContextKey('alwaysTrue', true);
-
-        ed.addCommand(monaco.KeyCode.Enter | monaco.KeyMod.Shift,function (e) {
-          let code = ed.getModel().getValue(monaco.editor.EndOfLinePreference.LF)
-          // let codeNode = <Graph.JsCodeNode> cell.code;
-          // codeNode.source = code;
-          // console.log(cell);
-          context.rebindSubsequent(cell, code)
-        }, 'alwaysTrue');
-
-        let lastHeight = 100;
-        let lastWidth = 0
-        let resizeEditor = () => {
-          let lines = ed.getModel().getValue(monaco.editor.EndOfLinePreference.LF, false).split('\n').length
-          let height = lines > 4 ? lines * 20.0 : 80;
-          let width = el.clientWidth
-
-          if (height !== lastHeight || width !== lastWidth) {
-            lastHeight = height
-            lastWidth = width  
-            ed.layout({width:width, height:height})
-            el.style.height = height + "px"
-          }
-        }
-        ed.getModel().onDidChangeContent(resizeEditor);
-        window.addEventListener("resize", resizeEditor)
-        setTimeout(resizeEditor, 100)
-      }
-      let code = h('div', { style: "height:100px; margin:20px 0px 10px 0px;", id: "editor_" + cell.editor.id.toString(), afterCreate:afterCreateHandler }, [ ])
+			let code = createEditor("javascript", state.block.source, cell, context)
       return h('div', { }, [code, results])
     },
   }
